bugfix: 1) ошибка с строке когда в запросе только бренд и город 2) ошибка где отправляет только один запрос из за чего данные не валидны 3) лишний пробел в поле payload
+ добавил unit тесты
