+++ requirements.txt
@@ -4,3 +4,4 @@ aiohttp_cors
 pyyaml
 asyncpg~=0.22.0
 aiohttp-requests
+asynctest~=0.13.0
